Add TestUtil component and move RenderWithRouter
