+++ src/components/index.js
@@ -1,5 +1,4 @@
 export { default as Navigation } from './Navigation';
-export { default as RenderWithRouter } from './RenderWithRouter';
 export { default as Button } from './Button';
 export { default as Logo } from './Logo';
 export { default as Container } from './Container';
