+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as TestUtil } from './TestUtil';
+export { default as RenderWithRouter } from './RenderWithRouter';
